+++ libs/core/src/lib/input-group/input-group.component.ts
@@ -145,6 +145,12 @@ export class InputGroupComponent implements ControlValueAccessor {
     }
 
     /** @hidden */
+    setDisabledState(isDisabled: boolean): void {
+        this.disabled = isDisabled;
+        this.changeDetectorRef.markForCheck();
+    }
+
+    /** @hidden */
     buttonClicked($event) {
         this.addOnButtonClicked.emit($event);
     }